pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cloudrun-langgraph-demo"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
keywords = []
authors = [
  { name = "Xiang Shen", email = "shenxiang@google.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "google-cloud-alloydb-connector>=1.7.0",
    "google-cloud-storage[tracing]>=2.19.0",
    "grpcio>=1.71.0",
    "langchain-community>=0.3.20",
    "langchain-core>=0.3.49",
    "langchain-google-vertexai>=2.0.18",
    "langgraph>=0.3.21",
    "opentelemetry-distro>=0.52b1",
    "opentelemetry-exporter-gcp-logging>=1.9.0a0",
    "opentelemetry-exporter-gcp-trace>=1.9.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.31.1",
    "opentelemetry-instrumentation-aiohttp-client>=0.52b1",
    "opentelemetry-instrumentation-httpx>=0.52b1",
    "opentelemetry-instrumentation-requests>=0.52b1",
    "opentelemetry-instrumentation-sqlalchemy>=0.52b1",
    "opentelemetry-instrumentation-sqlite3>=0.52b1",
    "opentelemetry-instrumentation-urllib>=0.52b1",
    "opentelemetry-instrumentation-urllib3>=0.52b1",
    "opentelemetry-instrumentation-vertexai>=2.0b0",
    "streamlit>=1.44.0",
]

[project.urls]
Documentation = "https://github.com/xiangshen-dk/cloudrun-langgraph-demo#readme"
Issues = "https://github.com/xiangshen-dk/cloudrun-langgraph-demo/issues"
Source = "https://github.com/xiangshen-dk/cloudrun-langgraph-demo"

[tool.hatch.version]
path = "src/cloudrun_langgraph_demo/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/cloudrun_langgraph_demo tests}"

[tool.coverage.run]
source_pkgs = ["cloudrun_langgraph_demo", "tests"]
branch = true
parallel = true
omit = [
  "src/cloudrun_langgraph_demo/__about__.py",
]

[tool.coverage.paths]
cloudrun_langgraph_demo = ["src/cloudrun_langgraph_demo", "*/cloudrun-langgraph-demo/src/cloudrun_langgraph_demo"]
tests = ["tests", "*/cloudrun-langgraph-demo/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[project.entry-points.opentelemetry_traces_exporter]
otlp_google_auth = "cloudrun_langgraph_demo._otlp_google_auth:OTLPSpanExporterGoogleAuth"
